% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data4Exam}
\alias{Data4Exam}
\title{Full simulation of an insurance market}
\format{tibble}
\usage{
Data4Exam
}
\description{
The complete dataset obtain through 100 simulation
}
\keyword{datasets}
