% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{DatasetEx}
\alias{DatasetEx}
\title{Average results of an insurance market}
\format{tibble}
\usage{
DatasetEx
}
\description{
The dataset obtain through means of all the variables presents in Data4Exam
}
\keyword{datasets}
