% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Org.group}
\alias{Org.group}
\title{A list of the organized data}
\format{tibble}
\usage{
Org.group
}
\description{
The nested list used to plot by group the data from DatasetEx
}
\keyword{datasets}
