% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Regmaker}
\alias{Regmaker}
\title{Regmaker}
\usage{
Regmaker()
}
\description{
This function automatically compound all the possible multilinear regression, with anova, using all variables present in the dataset and
gives back a rda file composed of all the information necessary organize
}
\details{
' @examples
\dontrun{
 Regmaker()
}
}
