% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{RegmakerCoeff}
\alias{RegmakerCoeff}
\title{Regmakercoeff}
\usage{
RegmakerCoeff()
}
\value{
data.frame
\describe{
  \item{Coeff_frame}{The output is a data.frame composed of all the coefficients and their relative P-values organize. }
   }
}
\description{
This function automatically compound all the possible multilinear regression, with anova, using all variables present in the dataset and
gives back a rda file composed of all the information necessary organize
}
\examples{

RegmakerCoeff()

}
