% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Plot.group}
\alias{Plot.group}
\title{Plot.group}
\usage{
Plot.group(n)
}
\arguments{
\item{n}{(1:4) 1 for grouping according to price variables, 2 revenues, 3 market share, 4 cartel's market share}
}
\description{
Return a linear plot of all 10 observed agent group according to n.
}
