% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\name{Plot.group}
\alias{Plot.group}
\title{Plot.group}
\usage{
Plot.group(n)
}
\arguments{
\item{n}{(1:4) 1 for grouping according to price variables, 2 revenues, 3 market share, 4 cartel's market share}
}
\value{
Plot
\describe{
  \item{Time}{Variable always used on the x-axis }
  \item{Price series}{Variable used on the y-axis if n=1, prices of the 10 observed firms}
  \item{Revenues series}{Variable used on the y-axis if n=2 Revenues of the 10 observed firms}
   \item{Market quotas}{Variable used on the y-axis if n=3 Market share of the 10 observed firms}
   \item{Cartel's market share}{Variable used on the y-axis if n=4 Cartel's market share of the 10 observed firms}
   }
}
\description{
Return a linear plot of all 10 observed agent group according to n.
}
\examples{
\dontrun{
Plot.group(1)

Plot.group(4)
}
}
